refac(env): rename env to secret
